requirements.txt
bcrypt
datetime
pyotp
mysql-connector-python